Rename ValuDeclaration to VariableDeclaration
